+++ packages/app/e2e/command-center-host.spec.ts
@@ -3,6 +3,7 @@ import { expect } from "@playwright/test";
 import { test } from "./fixtures";
 import { gotoAppShell } from "./helpers/app";
 import { createIdleAgent } from "./helpers/archive-tab";
+import { openCommandCenter } from "./helpers/command-center";
 import { appendOfflineHost } from "./helpers/hosts";
 import { seedWorkspace } from "./helpers/seed-client";
 import { getServerId } from "./helpers/server-id";
@@ -34,9 +35,7 @@ test.describe("Command center host labels", () => {
       });
       await gotoAppShell(page);
 
-      await page.getByTestId("sidebar-command-center-search").click();
-      const panel = page.getByTestId("command-center-panel");
-      await expect(panel).toBeVisible({ timeout: 30_000 });
+      const panel = await openCommandCenter(page);
 
       // The shared daemon may carry agents from other specs, so target this agent by its id.
       const row = panel.getByTestId(`command-center-agent-${getServerId()}:${agent.id}`);
