+++ packages/app/e2e/sidebar-host-filter-multi.spec.ts
@@ -1,7 +1,13 @@
 import { expect } from "@playwright/test";
 import { test } from "./fixtures";
 import { gotoAppShell } from "./helpers/app";
-import { appendOfflineHost } from "./helpers/hosts";
+import {
+  appendOfflineHost,
+  expectHostFilterRow,
+  openSidebarDisplayPreferences,
+  selectAllHostsFilter,
+  toggleHostFilter,
+} from "./helpers/hosts";
 import { seedWorkspace } from "./helpers/seed-client";
 import { getServerId } from "./helpers/server-id";
 
@@ -21,38 +27,27 @@ test.describe("Sidebar host filter (multi-select)", () => {
       // A second (offline) host is enough to surface the host filter without a second daemon.
       await appendOfflineHost(page, { serverId: SECONDARY_HOST_ID, label: "Secondary Host" });
       await gotoAppShell(page);
-
       await expect(workspaceRow).toBeVisible({ timeout: 30_000 });
 
-      await page.getByTestId("sidebar-display-preferences-menu").click();
-      await expect(page.getByTestId("sidebar-display-preferences-content")).toBeVisible({
-        timeout: 10_000,
-      });
-
-      // The filter section lists "All hosts" plus a row per host, each with a status dot on the left.
-      await expect(page.getByTestId("sidebar-host-filter-all")).toBeVisible();
-      await expect(page.getByTestId(`sidebar-host-filter-${serverId}`)).toBeVisible();
-      await expect(page.getByTestId(`sidebar-host-filter-${SECONDARY_HOST_ID}`)).toBeVisible();
-      await expect(page.getByTestId(`sidebar-host-filter-status-${serverId}`)).toBeVisible();
-      await expect(
-        page.getByTestId(`sidebar-host-filter-status-${SECONDARY_HOST_ID}`),
-      ).toBeVisible();
+      await openSidebarDisplayPreferences(page);
+      await expectHostFilterRow(page, serverId);
+      await expectHostFilterRow(page, SECONDARY_HOST_ID);
 
       // Pin the primary host — its workspace stays visible.
-      await page.getByTestId(`sidebar-host-filter-${serverId}`).click();
+      await toggleHostFilter(page, serverId);
       await expect(workspaceRow).toBeVisible();
 
       // Add the secondary host without clearing the primary. Under single-select this would replace
       // the primary and hide the workspace; multi-select keeps both pinned, so it stays visible.
-      await page.getByTestId(`sidebar-host-filter-${SECONDARY_HOST_ID}`).click();
+      await toggleHostFilter(page, SECONDARY_HOST_ID);
       await expect(workspaceRow).toBeVisible();
 
       // Drop the primary host — only the (empty) secondary host remains pinned, so the workspace hides.
-      await page.getByTestId(`sidebar-host-filter-${serverId}`).click();
+      await toggleHostFilter(page, serverId);
       await expect(workspaceRow).toHaveCount(0, { timeout: 10_000 });
 
       // Back to all hosts — the workspace returns.
-      await page.getByTestId("sidebar-host-filter-all").click();
+      await selectAllHostsFilter(page);
       await expect(workspaceRow).toBeVisible({ timeout: 10_000 });
     } finally {
       await seeded.cleanup();
